// clients/evm_repl/src/app/components/root/model/AppRootActions.tsx
enum AppRootActionType {
  FETCH_HOME = "FETCH_HOME",
  INPUT_CHANGED = "INPUT_CHANGED",
  EMULATOR_STATE_CHANGED = "EMULATOR_STATE_CHANGED",
  START_LOADING = "START_LOADING",
  COMMAND_EOSIO_API = "COMMAND_EOSIO_API",
  COMMAND_EVM_CONTRACT_NAME = "COMMAND_EVM_CONTRACT_NAME",
  COMMAND_EOS_ACCOUNT_NAME = "COMMAND_EOS_ACCOUNT_NAME",
  COMMAND_EOS_PRIVATE_KEY = "COMMAND_EOS_PRIVATE_KEY",
  COMMAND_CREATE = "COMMAND_CREATE",
  COMMAND_RAW = "COMMAND_RAW",
  COMMAND_ACCOUNT = "COMMAND_ACCOUNT",
  COMMAND_CODE = "COMMAND_CODE",
  COMMAND_STATE = "COMMAND_STATE",
}

export { AppRootActionType };